Older FFmpeg/LibAV versions still define unsigned error codes.
Amends: r2743

git-svn-id: svn://svn.daper.net/moc/trunk@2785 910807d9-36e0-0310-a014-e9ea483e2ba4

+++ decoder_plugins/ffmpeg/ffmpeg.c
@@ -1080,7 +1080,7 @@ static AVPacket *get_packet (struct ffmpeg_data *data)
 	free_packet (pkt);
 
 	/* FFmpeg has (at least) two ways of indicating EOF.  (Awesome!) */
-	if (rc == AVERROR_EOF)
+	if (rc == (int)AVERROR_EOF)
 		data->eof = true;
 	if (data->ic->pb && data->ic->pb->eof_reached)
 		data->eof = true;
